drf_api/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.6
# idna==3.4
# pip==21.3.1
# pytz==2023.3
# requests==2.31.0
# setuptools==60.2.0
# sqlparse==0.4.4
# tzdata==2023.3
# urllib3==2.0.2
# wheel==0.37.1
